+++ orcid-core/src/main/java/org/orcid/core/manager/v3/impl/MembersManagerImpl.java
@@ -22,6 +22,7 @@ import org.orcid.core.manager.v3.MembersManager;
 import org.orcid.core.manager.v3.ProfileEntityManager;
 import org.orcid.core.manager.v3.RecordNameManager;
 import org.orcid.core.manager.v3.SourceManager;
+import org.orcid.core.manager.v3.read_only.ClientDetailsManagerReadOnly;
 import org.orcid.core.manager.v3.read_only.ClientManagerReadOnly;
 import org.orcid.core.security.OrcidRoles;
 import org.orcid.jaxb.model.clientgroup.ClientType;
@@ -75,6 +76,9 @@ public class MembersManagerImpl implements MembersManager {
 
     @Resource(name = "clientManagerReadOnlyV3")
     private ClientManagerReadOnly clientManagerReadOnly;
+
+    @Resource(name = "clientDetailsManagerReadOnlyV3")
+    private ClientDetailsManagerReadOnly clientDetailsManagerReadOnly;
 
     @Resource
     private OrcidGenerationManager orcidGenerationManager;
@@ -277,7 +281,11 @@ public class MembersManagerImpl implements MembersManager {
                     Set<Client> clients = clientManagerReadOnly.getClients(orcid);
                     List<org.orcid.pojo.ajaxForm.Client> clientsList = new ArrayList<org.orcid.pojo.ajaxForm.Client>();
                     clients.forEach(c -> {
-                        clientsList.add(org.orcid.pojo.ajaxForm.Client.fromModelObject(c));
+                        org.orcid.pojo.ajaxForm.Client client = org.orcid.pojo.ajaxForm.Client.fromModelObject(c);
+                        ClientDetailsEntity clientDetails = clientDetailsManagerReadOnly.findByClientId(c.getId());
+                        boolean isDeactivated = clientDetails.getDeactivatedDate() != null;
+                        client.setDeactivated(isDeactivated);
+                        clientsList.add(client);
                     });
                     member.setClients(clientsList);
                 } else {