+++ orcid-web/src/test/java/org/orcid/frontend/web/controllers/WorksControllerTest.java
@@ -172,12 +172,12 @@ public class WorksControllerTest extends BaseControllerTest {
 
         contributor = work.getContributorsGroupedByOrcid().get(2);
         assertNull(contributor.getContributorEmail());
-        assertEquals("Not This Name", contributor.getCreditName().getContent());
+        assertEquals("Credit Name", contributor.getCreditName().getContent());
         
         // contributor is an ORCID user with private name
         contributor = work.getContributorsGroupedByOrcid().get(3);
         assertNull(contributor.getContributorEmail());
-        assertEquals("Not This Name", contributor.getCreditName().getContent());
+        assertEquals("Name is private", contributor.getCreditName().getContent());
     }
 
     @Test
@@ -197,11 +197,11 @@ public class WorksControllerTest extends BaseControllerTest {
 
         contributor = work.getContributorsGroupedByOrcid().get(2);
         assertNull(contributor.getContributorEmail());
-        assertEquals("Not This Name", contributor.getCreditName().getContent());
+        assertEquals("Credit Name", contributor.getCreditName().getContent());
 
         contributor = work.getContributorsGroupedByOrcid().get(3);
         assertNull(contributor.getContributorEmail());
-        assertEquals("Not This Name", contributor.getCreditName().getContent());
+        assertEquals("Name is private", contributor.getCreditName().getContent());
     }
 
     @Test