+++ CHANGELOG.md
@@ -1,3 +1,65 @@
+## v2.96.0 - 2025-06-18
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.95.0...v2.96.0)
+
+### Fix
+
+- include client deactivated status in findMember result (#7321)
+
+## v2.95.0 - 2025-06-17
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.94.0...v2.95.0)
+
+### Feature
+
+- Add deactivated status in findClient API response (#7318)
+
+## v2.94.0 - 2025-06-17
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.93.0...v2.94.0)
+
+## v2.93.0 - 2025-06-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.92.0...v2.93.0)
+
+## v2.92.0 - 2025-06-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.91.0...v2.92.0)
+
+## v2.91.0 - 2025-06-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.90.0...v2.91.0)
+
+## v2.90.0 - 2025-06-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.89.0...v2.90.0)
+
+## v2.89.0 - 2025-06-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.88.0...v2.89.0)
+
+## v2.88.0 - 2025-06-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.87.0...v2.88.0)
+
+## v2.87.0 - 2025-06-11
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.86.0...v2.87.0)
+
+## v2.86.0 - 2025-06-09
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.85.0...v2.86.0)
+
+## v2.85.0 - 2025-06-06
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.84.0...v2.85.0)
+
+- [#7310](https://github.com/ORCID/ORCID-Source/pull/7310): add sast ci
+
+## v2.84.0 - 2025-06-05
+
+[Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.83.0...v2.84.0)
+
 ## v2.83.0 - 2025-06-03
 
 [Full Changelog](https://github.com/ORCID/ORCID-Source/compare/v2.82.0...v2.83.0)